plugins/aws-agentcore-langgraph/examples/langgraph-web-search/pyproject.toml
[project]
name = "langgraph-web-search"
version = "0.1.0"
description = "LangGraph web search agent for AWS Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.0.7",
    "ddgs>=8.0.0",
    "langchain>=1.1.0",
    "langchain-aws>=1.1.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.0",
]
